impl: rework the URI handler
With this commit we split the responsibilities. `CoderProtocolHander` is now only
responsible searching the workspace and agent and orchestrating the IDE installation and launching.
The code around initializing the http client and cli with a new URL and a new token, plus cleaning
up the old resources like the polling loop and the list of environments.
There are two major benefits to this approach:
- allows us to easily share/reuse the logic around cleaning up resources and re-initializing the
http client and cli without passing so many callbacks to CoderProtocolHandler (less coupling, code that
is cleaner and easier to read, easier to maintain and test)
- provides a nice and easy way to check whether the URI url is the same as the one from current
deployment and properly react if they differ.