impl: improve URI flow when Toolbox is closed
When Toolbox is closed and URI is executed the handler waits for the
plugin to fully initialize with the last successful deployment after
which it goes on with the URI handling logic.
We can improve this situation by skipping the last successful
deployment which anyway. This reduces a lot the time to an actual IDE connection.