impl: model for the IDE product
This is the IDE model from the https://data.services.jetbrains.com/products? feed.
Only `release` and `eap` IDEs will be supported for now.

This commit is in preparation for supporting placeholders like `last_release` or `last_eap`
in the URI handler. Unfortunately Toolbox does not expose a way to determine if a build like
251.2829.367 is an eap build or an actual release, so we have to look up the metadata in the
feed mentioned earlier.