impl: load and cache IDEs
This commit implements an IDE feed manager that depending on whether Toolbox runs in offline mode or
not loads the eap and release information from two local json files or from the https://data.services.jetbrains.com/products.
The feed manager is also able to intersect a list of available build versions (the IDE versions available for install
on a workspace) with the list of versions fetched from the feed and return the latest version available for install.
The intersection happens only after the feed was filtered by product code release type.

Support for offline mode is WIP. The commit assumes that in offline mode Toolbox exposes a certain system property.