impl: support for `latest_eap`, `latest_release` and `latest_installed` placeholders
This PR adds support for the IDE placeholders in the URI build number parameter. This is a request
coming from Netflix, the placeholders provide an easier way to fill in the build number from the
web dashboard without having to know the available versions of IDE from Toolbox.

IDE launch flow is refactored to support dynamic build selectors, and it also improves the
install/launch logic. When latest_eap or latest_release is used:
- prefer the newest matching version that is available for install
- install it only if it isn’t already installed
- fall back to the latest available version if no match exists
- show an error if none of the above happens

When latest_installed is used:
- launch the newest installed version
- if none are installed, install and launch the latest available version
- show an error if none of the above happens