fix: mTLS certificate refresh
A mechanism to intercept expired certificate errors was
introduced in previous releases. Upon intercept the a
command was spawend to refresh the certs and then the request
was retried again.

Unfortunately the interceptor mechanism had a big drawback, it only
acts on an invalid response from the server. Most of the times the
SSL handshake exceptions are raised much earlier, when the connection
to the server is established. In this place the interceptor does
not work.

A different approach was instead used. All http calls are now wrapped
in an executor that is able to retry the request if there is an
exception in any of the request lifecycle.