fix: don't create new environment instances every time we poll
The polling logic has a list of "environments" which is updated by the poll loop.
The loop retrieves a list of workspaces and agents from the backend, and then it
looks if the pairs are already part of the existing environments. This is where
we have an issue: we look by temporarily creating a new "RemoteEnvironment" and use
the id of the temporary env. to search in the existing collection. Besides creating
a lot of temporary objects every 5 seconds, the logic has another flaw: the constructor
of the RemoteEnvironment triggers the ssh connection for that workspace if it was marked
in the settings store as connected in the previous state. Asking to establish the ssh
connection every 5 seconds is completely unnecessary, from my testings the newer versions
of Toolbox are able to remember in the same session that a connection was opened, and it
needs to be re-established again if it fails during that session.

From now on the constructor for the RemoteEnvironment will request it only for new login
sessions, from then on it is up to TBX to remember the connection state.
In the early versions we added custom logic to retrigger the ssh connection when API keys
expired. Turns out that is also no longer necessary to do. When an API key expires the
user will have to go again through the login sequence, which means starting with a clean
sheet of "environments". At the first poll run the RemoteEnvironment constructor will request
TBX to establish the ssh connection again.